Small bug fix in encoding.
The root of each data structure was wrongly asserted in B to be among
the locations x_1,...,x_n, which does not hold for empty lists rooted
in null. Formulas such as not list(null) were thus wrongly deemed
unsatisfiable.

Replaced the above constraint by an assertion that null is not
reachable from itself, which does not follow from the axiomatization
of the reachability predicates. This ensures that list(x) /\ not
list(x) is unsatisfiable even after removing the above assertion.